refactor: update env example, rename SponsorSection, redirect root page to home
