+++ .env.example
@@ -15,9 +15,29 @@
 # https://next-auth.js.org/configuration/options#secret
 AUTH_SECRET=""
 
-# Prisma
-# https://www.prisma.io/docs/reference/database-reference/connection-urls#env
-DATABASE_URL="postgresql://postgres:password@localhost:5432/oddsnends"
-
 # LINE
 NEXT_PUBLIC_LIFF_ID=""
+
+BLOB_READ_WRITE_TOKEN=""
+
+# Recommended for most uses
+DATABASE_URL=
+
+# For uses requiring a connection without pgbouncer
+DATABASE_URL_UNPOOLED=
+
+# Parameters for constructing your own connection string
+PGHOST=
+PGHOST_UNPOOLED=
+PGUSER=
+PGDATABASE=
+PGPASSWORD=
+
+# Parameters for Vercel Postgres Templates
+POSTGRES_URL=
+POSTGRES_URL_NON_POOLING=
+POSTGRES_HOST=
+POSTGRES_PASSWORD=
+POSTGRES_DATABASE=
+POSTGRES_URL_NO_SSL=
+POSTGRES_PRISMA_URL=
