+++ src/components/SocialMediaBar/SocialMediaBar.tsx
@@ -1,34 +1,34 @@
-import SponsorSection from "../SponserSection/SponsorSection";
+import SponsorSection from "../SponsorSection/SponsorSection";
 import SocialMediaLink from "./SocialMediaLink/SocialMediaLink"; // Social media Link component
 
 function SocialMediaBar() {
-    return (
-        <>
-            <SponsorSection />
-            <footer className="bg-gray-200 text-black w-full px-4 py-4 sticky bottom-0 z-10">
-                <div className="flex items-center justify-around w-full">
-                    <div className="flex items-center justify-center gap-2">
-                        <SocialMediaLink
-                            href="https://www.instagram.com/oddsnends.fest/"
-                            src="/images/instagram.png"
-                            alt="Instagram"
-                            title="Instagram"
-                            label="oddsnends.fest"
-                        />
-                    </div>
-                    <div className="flex items-center justify-center gap-2">
-                        <SocialMediaLink
-                            href="https://www.tiktok.com/@oddsnends.fest"
-                            src="/images/tik-tok.png"
-                            alt="TikTok"
-                            title="TikTok"
-                            label="oddsnends.fest"
-                        />
-                    </div>
-                </div>
-            </footer>
-        </>
-    )
+  return (
+    <>
+      <SponsorSection />
+      <footer className="bottom-0 z-10 w-full bg-gray-200 px-4 py-4 text-black">
+        <div className="flex w-full items-center justify-around">
+          <div className="flex items-center justify-center gap-2">
+            <SocialMediaLink
+              href="https://www.instagram.com/oddsnends.fest/"
+              src="/images/instagram.png"
+              alt="Instagram"
+              title="Instagram"
+              label="oddsnends.fest"
+            />
+          </div>
+          <div className="flex items-center justify-center gap-2">
+            <SocialMediaLink
+              href="https://www.tiktok.com/@oddsnends.fest"
+              src="/images/tik-tok.png"
+              alt="TikTok"
+              title="TikTok"
+              label="oddsnends.fest"
+            />
+          </div>
+        </div>
+      </footer>
+    </>
+  );
 }
 
 export default SocialMediaBar;
