+++ src/components/SponsorSection/SponsorSection.tsx
